+++ docs/change-log.md
@@ -6,6 +6,7 @@ outline: deep
 ## 2024-04-04
 * Provided a little more documentation about virtual columns
 * Fixed camelCasing of schema names in the schema generator
+* Added mediumtext fields to schema and drizzle generation
 
 ## 2024-03-16
 * Created a function for fetching ving config in `ving/config.mjs`